Backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.3
pydantic==2.9.1
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
